Fix bug where after leaving/getting kicked from room could not re-join
This was caused by some trippy stuff in useEffect and room data getting pulled in unexpectedly because that page useEffect gets called all the time. I should rewrite that when I figure out better how it works.

How I have it now is it will only add room state to the slice if we have a player id. This means that when a player leaves/is kicked, the player id is cleared and then when the room_change socket callback tries to update the room state, it isn't allowed to.